Prevent store attendant from deleting a product

+++ api/views.py
@@ -276,6 +276,15 @@ class ProductView(MethodView):
         """
         db_conn = DB()
 
+        # Get logged in user
+        current_user = get_jwt_identity()
+        loggedin_user = db_conn.get_user(current_user)
+        # # Check if it's not store owner
+        if not loggedin_user["is_admin"]:
+            return jsonify({
+                "error": "Please login as a store owner"
+            }), 403
+
         # Delete product
         db_conn.delete_product(int(product_id))
         return jsonify({
